Make abstract VideoAsyncReader class for TensorFlow based detectors
This allows the async feature to be used with both TensorFlow and
TensorNets object detectors.
